board: moved boardIndex into separated file
