+++ ttgboard/board.go
@@ -50,43 +50,6 @@ func (p PlayerType) String() string {
 	return "?"
 }
 
-// BoardIndex represents board index
-type BoardIndex struct {
-	state IdxState
-}
-
-func newIndex() *BoardIndex {
-	result := &BoardIndex{
-		state: IdxNone,
-	}
-
-	return result
-}
-
-// SetState sets index state
-func (b *BoardIndex) SetState(state IdxState) {
-	b.state = state
-}
-
-func (b *BoardIndex) String() string {
-	switch b.state {
-	case IdxNone:
-		return " "
-	case IdxX:
-		return "X"
-	case IdxO:
-		return "O"
-	}
-
-	// should not be reached
-	return "?"
-}
-
-// IsFree return's true if index is free
-func (b BoardIndex) IsFree() bool {
-	return b.state == IdxNone
-}
-
 type player struct {
 	name       string
 	playerType PlayerType
